Use uint32 for chainID in chain registration
These functions were taking a mix of uint16 and uint32 for the chain id,
which is just a bug waiting to happen.  Use uint32 for all functions
because that's what the protobuf uses.

The chain id is still encoded as a uint16 in the various payloads but
it's trivial to cast a uint16 to a uint32 whereas we would need to
insert a bounds check every time we go the other way.

commit-id:42e6fe36

+++ wormhole_chain/x/tokenbridge/keeper/grpc_query_chain_registration.go
@@ -52,7 +52,7 @@ func (k Keeper) ChainRegistration(c context.Context, req *types.QueryGetChainReg
 
 	val, found := k.GetChainRegistration(
 		ctx,
-		uint16(req.ChainID),
+		req.ChainID,
 	)
 	if !found {
 		return nil, status.Error(codes.InvalidArgument, "not found")
